Fixed multiname bugs in AbcConstantPool, Refactored name interfaces, general cleanup
In AbcConstantPool, multiname() would return 0 if the argument's name was '*', which
is incorrect. Instead, it checks for equality against QName("*")

An interface IName was added, which extends IHashable. All names now implement IName,
or IMultiname (which extends IName). This should be more flexible.

An index of 0 is undefined for namespace and namespacesset, so those functions can
not return such anymore.